Rename ClassTool to AbstractClassTool and CustomTool to ClassTool and PredefinedTool to ToolboxTool
